devd: embed .hh file
